Tests, rename ApiBase ApiAbtract
