+++ pyairtable/api/base.py
@@ -1,9 +1,9 @@
 from typing import List
 
-from .api import ApiBase
+from .abstract import ApiAbstract
 
 
-class Base(ApiBase):
+class Base(ApiAbstract):
     """
     Represents an Airtable Base. This calss is similar to :class:`~pyairtable.api.Api`,
     except ``base_id`` is provided on init instead of provided on each method call.
